+++ Campaign/app/Http/Resources/SearchResource.php
@@ -17,10 +17,12 @@ class SearchResource extends JsonResource
     {
         $banners = $this->get('banners');
         $campaigns = $this->get('campaigns');
+        $snippets = $this->get('snippets');
 
         return Arr::collapse([
             $this->when($banners->isNotEmpty(), BannerSearchResource::collection($banners)->toArray(app('response'))),
-            $this->when($campaigns->isNotEmpty(), CampaignSearchResource::collection($campaigns)->toArray(app('response')))
+            $this->when($campaigns->isNotEmpty(), CampaignSearchResource::collection($campaigns)->toArray(app('response'))),
+            $this->when($snippets->isNotEmpty(), SnippetSearchResource::collection($snippets)->toArray(app('response'))),
         ]);
     }
 }